+++ kernel/ramdisk.h
@@ -3,4 +3,4 @@
 #include <stdbool.h>
 
 bool ramdisk_load_from_path(const char* path);
-bool ramdisk_load_from_module(uint32_t start, uint32_t end, const char* name);
+bool ramdisk_load_from_module(uintptr_t start, uintptr_t end, const char* name);
